chore:work on CRUD operations for bucketlist and items

+++ src/app/bucketlists/bucketlists.component.ts
@@ -13,7 +13,7 @@ import { BucketlistsService } from '../services/bucketlists.service'
 })
 export class BucketlistsComponent implements OnInit {
 
-  placeholder:string = "E.g Travel";
+  placeholder:string = 'E.g Travel';
   bucket_res: any;
   currentuser: string;
   buckets: any;
@@ -22,7 +22,7 @@ export class BucketlistsComponent implements OnInit {
   pages: 0;
   per_page: 0;
   add_bucket_res: any;
-  bucket ={"name": "Bucket Name"}
+  bucket ={'name': 'Bucket Name'}
   edit_bucket_name: any;
   bucket_id_to_edit: any;
   constructor(private authService: AuthService,
@@ -33,20 +33,20 @@ export class BucketlistsComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.currentuser = localStorage.getItem("currentuser");
+    this.currentuser = localStorage.getItem('currentuser');
   	this.getBucketlists();
 
   }
 
   getBucketlists(){
-    console.log("before request");
+    console.log('before request');
     this.bucketlistsService.getBuckets().subscribe((res: Response)=>{
       this.bucket_res = res.json();
       this.buckets = this.bucket_res.items;
       this.page = this.bucket_res.page;
       this.total_buckets = this.bucket_res.total;
       this.pages = this.bucket_res.pages;
-      console.log("In the request response");
+      console.log('In the request response');
       console.log(this.bucket_res)
       console.log('After request response');
     })
@@ -60,10 +60,10 @@ export class BucketlistsComponent implements OnInit {
   addBucket(f: NgForm): void {
     this.bucket.name =f.value.name
     console.log(f.value.name)
-    console.log("add bucket starts");
+    console.log('add bucket starts');
     this.bucketlistsService.postBucket(this.bucket).subscribe((res: Response)=>{
         this.add_bucket_res = res.json();
-        console.log("In add bucket")
+        console.log('In add bucket')
         console.log(this.add_bucket_res)
         console.log(this.add_bucket_res);
         this.getBucketlists();
@@ -71,37 +71,42 @@ export class BucketlistsComponent implements OnInit {
 
   }
 
-  deleteBucket(id){
-      this.bucketlistsService.deleteBucket(id).subscribe((res:Response)=>{
-      console.log("succesfly deleted item");
-      this.getBucketlists();
-
-    });
+  deleteBucket(id) {
+    let verify: boolean = confirm(`Are you sure you want to delete this bucket?`);
+    if (verify === true) {
+      this.bucketlistsService.deleteBucket(id).subscribe((res: Response) => {
+        console.log('succesfly deleted item');
+        this.getBucketlists();
+   });
+    }
   }
-  setBucketNameEditVariables(id, name){
-    console.log("start setting variables")
-    this.edit_bucket_name = name
-    console.log(this.edit_bucket_name)
-    this.bucket_id_to_edit = id
-    console.log(this.bucket_id_to_edit)
-    console.log("end setting variables")
+  setBucketNameEditVariables(id, name) {
+    console.log('start setting variables');
+    this.edit_bucket_name = name;
+    console.log(this.edit_bucket_name);
+    this.bucket_id_to_edit = id;
+    console.log(this.bucket_id_to_edit);
+    console.log('end setting variables');
   }
-  editBucketName(editform: NgForm): void{
-    console.log(this.edit_bucket_name)
-    this.edit_bucket_name = editform.value.name;
-    let name ={ "name": this.edit_bucket_name}
-    console.log(this.edit_bucket_name)
-    this.bucketlistsService.editBucket(this.bucket_id_to_edit, name).subscribe((res:Response)=>{
-      console.log("Bucket edited succesfully");
-      this.getBucketlists();
-    })
+  editBucketName(editform: NgForm): void {
+    let verify: boolean = confirm(`Are you sure you want to edit this bucket?`);
+    if (verify === true) {
+      console.log(this.edit_bucket_name);
+      this.edit_bucket_name = editform.value.name;
+      const name = { 'name': this.edit_bucket_name};
+      console.log(this.edit_bucket_name);
+      this.bucketlistsService.editBucket(this.bucket_id_to_edit, name).subscribe((res: Response) => {
+        console.log('Bucket edited succesfully');
+        this.getBucketlists();
+      });
+    }
   }
 
   viewBucket(bucketlist_id, bucketlist_name){
-    console.log("redirecting to bucketlists items");
-    localStorage.setItem("bucketlist_id", bucketlist_id);
-    localStorage.setItem("bucketlist_name",bucketlist_name);
-    console.log(localStorage.getItem("bucketlist_id"));
+    console.log('redirecting to bucketlists items');
+    localStorage.setItem('bucketlist_id', bucketlist_id);
+    localStorage.setItem('bucketlist_name',bucketlist_name);
+    console.log(localStorage.getItem('bucketlist_id'));
     this.router.navigate(['/items']);
   }
 }
